# docs/swagger.yaml
definitions:
  model.BaseResp:
    properties:
      code:
        description: http status at header response api-gateway
        type: integer
      data:
        type: object
      responseCode:
        type: string
      responseDesc:
        type: string
    type: object
  model.PosDetailProduct:
    properties:
      productId:
        type: string
      productName:
        type: string
      quantity:
        type: integer
      unitPrice:
        type: integer
    required:
    - productId
    - productName
    - quantity
    - unitPrice
    type: object
  model.ReqCheckStatus:
    properties:
      earningId:
        type: string
    required:
    - earningId
    type: object
  model.ReqEarningHandler:
    properties:
      accountNo:
        type: string
      earningId:
        type: string
      requestAt:
        type: string
    required:
    - accountNo
    - earningId
    - requestAt
    type: object
  model.ReqProductList:
    properties:
      invoiceNo:
        type: string
    required:
    - invoiceNo
    type: object
  model.ReqQrString:
    properties:
      detailProduct:
        items:
          $ref: '#/definitions/model.PosDetailProduct'
        type: array
      earningCode:
        type: string
      invoiceNo:
        type: string
      merchantId:
        type: string
      serviceCharge:
        type: integer
      staffId:
        type: string
      staffName:
        type: string
      storeId:
        type: string
      subTotalAmount:
        type: integer
      tax:
        type: integer
      totalAmount:
        type: integer
      transactionTime:
        type: string
    required:
    - detailProduct
    - invoiceNo
    - merchantId
    - staffId
    - staffName
    - storeId
    - subTotalAmount
    - totalAmount
    - transactionTime
    type: object
  model.ResQrString:
    properties:
      qrString:
        type: string
    type: object
info:
  contact:
    email: info@trading-ai.id
    name: Bumi Langit Team
    url: https://trading-ai.id
  license:
    name: Trading AI License
paths:
  /trading-ai/v1/earning/qrstring:
    post:
      consumes:
      - application/json
      operationId: ReqQrString
      parameters:
      - description: UUID Partner (example = 'PSM0002')
        in: header
        name: Institutionid
        required: true
        type: string
      - description: Hardcoded
        in: header
        name: Deviceid
        required: true
        type: string
      - description: ' '
        in: header
        name: Geolocation
        required: true
        type: string
      - description: Hardcoded (example = 'H2H')
        in: header
        name: Channelid
        required: true
        type: string
      - description: Generated system by partner, all request POST validate with signature
          (example = 'a0acd4cc5bb2')
        in: header
        name: Signature
        required: true
        type: string
      - description: (Example = 'pos-customer')
        in: header
        name: Appsid
        required: true
        type: string
      - description: ' '
        in: header
        name: Authorization
        required: true
        type: string
      - description: (Example = '1579666534')
        in: header
        name: Timestamp
        required: true
        type: string
      - description: Request Body
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/model.ReqQrString'
      produces:
      - application/json
      responses:
        "200":
          description: ' '
          schema:
            allOf:
            - $ref: '#/definitions/model.BaseResp'
            - properties:
                data:
                  $ref: '#/definitions/model.ResQrString'
              type: object
      summary: Request QR String from Hellobil
      tags:
      - Pos Service
  /trading-ai/v1/earning/request:
    post:
      consumes:
      - application/json
      operationId: ReqEarning
      parameters:
      - description: UUID Partner (example = 'PSM0002')
        in: header
        name: Institutionid
        required: true
        type: string
      - description: Hardcoded
        in: header
        name: Deviceid
        required: true
        type: string
      - description: ' '
        in: header
        name: Geolocation
        required: true
        type: string
      - description: Hardcoded (example = 'H2H')
        in: header
        name: Channelid
        required: true
        type: string
      - description: Generated system by partner, all request POST validate with signature
          (example = 'a0acd4cc5bb2')
        in: header
        name: Signature
        required: true
        type: string
      - description: (Example = 'pos-customer')
        in: header
        name: Appsid
        required: true
        type: string
      - description: ' '
        in: header
        name: Authorization
        required: true
        type: string
      - description: (Example = '1579666534')
        in: header
        name: Timestamp
        required: true
        type: string
      - description: Request Body
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/model.ReqEarningHandler'
      produces:
      - application/json
      responses:
        "200":
          description: ' '
          schema:
            allOf:
            - $ref: '#/definitions/model.BaseResp'
            - properties:
                data:
                  $ref: '#/definitions/model.ResQrString'
              type: object
      summary: hit from apps mobile / partner user apps
      tags:
      - Pos Service
  /trading-ai/v1/earning/status:
    post:
      consumes:
      - application/json
      operationId: CheckStatus
      parameters:
      - description: UUID Partner (example = 'PSM0002')
        in: header
        name: Institutionid
        required: true
        type: string
      - description: Hardcoded
        in: header
        name: Deviceid
        required: true
        type: string
      - description: ' '
        in: header
        name: Geolocation
        required: true
        type: string
      - description: Hardcoded (example = 'H2H')
        in: header
        name: Channelid
        required: true
        type: string
      - description: Generated system by partner, all request POST validate with signature
          (example = 'a0acd4cc5bb2')
        in: header
        name: Signature
        required: true
        type: string
      - description: (Example = 'pos-customer')
        in: header
        name: Appsid
        required: true
        type: string
      - description: ' '
        in: header
        name: Authorization
        required: true
        type: string
      - description: (Example = '1579666534')
        in: header
        name: Timestamp
        required: true
        type: string
      - description: ' '
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/model.ReqCheckStatus'
      produces:
      - application/json
      responses:
        "200":
          description: ' '
          schema:
            $ref: '#/definitions/model.BaseResp'
      summary: Check Status Hit by ottopoint-stamp
      tags:
      - Pos Service
  /trading-ai/v1/product/list:
    post:
      consumes:
      - application/json
      operationId: ProductList
      parameters:
      - description: UUID Partner (example = 'PSM0002')
        in: header
        name: Institutionid
        required: true
        type: string
      - description: Hardcoded
        in: header
        name: Deviceid
        required: true
        type: string
      - description: ' '
        in: header
        name: Geolocation
        required: true
        type: string
      - description: Hardcoded (example = 'H2H')
        in: header
        name: Channelid
        required: true
        type: string
      - description: Generated system by partner, all request POST validate with signature
          (example = 'a0acd4cc5bb2')
        in: header
        name: Signature
        required: true
        type: string
      - description: (Example = 'pos-customer')
        in: header
        name: Appsid
        required: true
        type: string
      - description: ' '
        in: header
        name: Authorization
        required: true
        type: string
      - description: (Example = '1579666534')
        in: header
        name: Timestamp
        required: true
        type: string
      - description: ' '
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/model.ReqProductList'
      produces:
      - application/json
      responses:
        "200":
          description: ' '
          schema:
            $ref: '#/definitions/model.BaseResp'
      summary: Product List Hit by admin portal
      tags:
      - Pos Service
swagger: "2.0"